requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
altair=4.1.0=py_1
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=py310h7516544_0
asttokens=2.2.1=pyhd8ed1ab_0
attrs=22.1.0=py310h06a4308_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py310h06a4308_0
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.5.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
contourpy=1.0.5=py310hdb19cb5_0
cryptography=39.0.1=py310h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py310h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py310h06a4308_0
glib=2.69.1=he621ea3_2
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
importlib-metadata=6.0.0=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.13.2=pyh41d4057_0
ipywidgets=8.0.4=py310h06a4308_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=py310h06a4308_0
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=py310h06a4308_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=4.12.0=py310hff52083_0
jupyterlab_widgets=3.0.5=py310h06a4308_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcurl=8.1.1=h91b91d3_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.52.0=ha637b67_1
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=hcc01f38_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.2.0=py310h06a4308_1
markupsafe=2.1.1=py310h7f8727e_0
matplotlib=3.7.1=py310h06a4308_1
matplotlib-base=3.7.1=py310h1128e8f_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.0=py310h06a4308_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py310h06a4308_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=py310h85018f9_1
numpy=1.24.3=py310h5f9d8c6_1
numpy-base=1.24.3=py310hb5e798b_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1u=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.0=py310h06a4308_0
pandas=2.0.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
ply=3.11=py310h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.38=pyha770c72_0
prompt_toolkit=3.0.38=hd8ed1ab_0
protobuf=3.20.3=py310h6a678d5_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py310h468efa6_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py310h06a4308_0
pyg=2.3.0=py310_torch_2.0.0_cu118
pygments=2.15.1=pyhd8ed1ab_0
pympler=0.9=py_0
pyopenssl=23.0.0=py310h06a4308_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.11=h7a1cb2a_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=2.0.1=py3.10_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0=py310h5eee18b_1
pyzmq=25.0.2=py310h6a678d5_0
qt-main=5.15.2=h8373d8f_8
qt-webengine=5.15.9=hbbf29b9_6
qtwebkit=5.212=h3fafdc1_5
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.29.0=py310h06a4308_0
rich=13.3.5=py310h06a4308_0
scikit-learn=1.2.2=py310h6a678d5_0
scipy=1.10.1=py310h5f9d8c6_1
semver=2.13.0=pyhd3eb1b0_0
setuptools=66.0.0=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyh6c4a22f_0
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.16.0=py310h06a4308_0
sympy=1.11.1=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py310h06a4308_0
torchaudio=2.0.2=py310_cu118
torchmetrics=0.11.2=py310h2f386ee_0
torchtriton=2.0.0=py310
torchvision=0.15.2=py310_cu118
tornado=6.1=py310h5764c6d_3
tqdm=4.65.0=py310h2f386ee_0
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.5.0=py310h06a4308_0
tzdata=2023.3=pypi_0
tzlocal=2.1=py310h06a4308_1
urllib3=1.26.15=py310h06a4308_0
utf8proc=2.6.1=h27cfd23_0
validators=0.18.2=pyhd3eb1b0_0
watchdog=2.1.6=py310h06a4308_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py310h06a4308_0
widgetsnbextension=4.0.5=py310h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.11.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
